fix(frontend): decide benger-vs-Vertretbar landing server-side (no flash, correct title)
The Next middleware /->/vertretbar rewrite never compiled into the image (middleware-manifest is empty; benger's subdomain routing is client-side too), so app/page.tsx SSR'd the benger homepage then client-swapped to VertretbarLanding — a visible flash on vertretbar.net. Make app/page.tsx a server component that resolves the host (x-forwarded-host) and renders the Vertretbar landing gate on student-locked hosts, benger landing otherwise; generateMetadata sets the Vertretbar <title>. Benger client landing moved verbatim to BengerLandingClient.tsx.
