+++ services/frontend/src/app/BengerLandingClient.tsx
@@ -1,0 +1,90 @@
+'use client'
+// BenGER Frontend v3.0.1 - Force rebuild for latest changes
+
+import { HeroSection } from '@/components/landing/HeroSection'
+import { InformationSection } from '@/components/landing/InformationSection'
+import { LicenseCitationSection } from '@/components/landing/LicenseCitationSection'
+import { LandingLayout } from '@/components/landing/LandingLayout'
+import { NewsSection } from '@/components/landing/NewsSection'
+import { PeopleSection } from '@/components/landing/PeopleSection'
+import { useAuth } from '@/contexts/AuthContext'
+import { useI18n } from '@/contexts/I18nContext'
+import { useSlot } from '@/lib/extensions/slots'
+import { isStudentLockedHost } from '@/lib/utils/subdomain'
+import { authRedirect } from '@/utils/authRedirect'
+import { useRouter } from 'next/navigation'
+import { useEffect, useState } from 'react'
+
+// The benger marketing landing. Rendered by app/page.tsx on non-student-locked
+// hosts (the host is decided server-side there). The client-side isStudentLockedHost
+// branch below is a dev/safety fallback — on vertretbar.net the server component
+// renders the Vertretbar landing directly, so this never runs there in prod.
+
+export default function BengerLandingClient() {
+  const { user, isLoading } = useAuth()
+  const { t } = useI18n()
+  const router = useRouter()
+  const VertretbarLanding = useSlot('VertretbarLanding')
+  const [isLocked, setIsLocked] = useState(false)
+
+  // SSR has no auth context — always render the landing layout server-side and
+  // wait for the client to mount before branching on isLoading / user. Without
+  // this, hydration sees the server's "not authenticated" tree but the client
+  // already knows the user is logged in (cookie present), causing a mismatch.
+  const [mounted, setMounted] = useState(false)
+  // eslint-disable-next-line react-hooks/set-state-in-effect
+  useEffect(() => { setMounted(true); setIsLocked(isStudentLockedHost()) }, [])
+
+  // Redirect authenticated users to dashboard
+  // Only redirect when auth state is stable (not loading)
+  useEffect(() => {
+    if (mounted && !isLoading && user) {
+      authRedirect.toDashboard(router)
+    }
+  }, [mounted, user, isLoading, router])
+
+  // Prevent flash while redirecting
+  if (mounted && !isLoading && user) {
+    return (
+      <div className="flex min-h-screen items-center justify-center bg-white dark:bg-zinc-900">
+        <div className="text-center">
+          <div className="mx-auto mb-4 h-8 w-8 animate-spin rounded-full border-b-2 border-emerald-500"></div>
+          <p className="text-zinc-600 dark:text-zinc-400">
+            {t('common.redirectingToDashboard')}
+          </p>
+        </div>
+      </div>
+    )
+  }
+
+  // Show loading state while auth is being checked (client-only)
+  if (mounted && isLoading) {
+    return (
+      <div className="flex min-h-screen items-center justify-center bg-white dark:bg-zinc-900">
+        <div className="text-center">
+          <div className="mx-auto mb-4 h-8 w-8 animate-spin rounded-full border-b-2 border-emerald-500"></div>
+          <p className="text-zinc-600 dark:text-zinc-400">{t('common.loading')}</p>
+        </div>
+      </div>
+    )
+  }
+
+  // Student-locked host (vertretbar.net), unauthenticated → the branded
+  // Vertretbar landing instead of the benger one (dev fallback for the
+  // server-side host branch in app/page.tsx). Authenticated users were already
+  // sent to /student above via authRedirect.toDashboard.
+  if (mounted && isLocked && !user && VertretbarLanding) {
+    // eslint-disable-next-line react-hooks/static-components
+    return <VertretbarLanding />
+  }
+
+  return (
+    <LandingLayout>
+      <HeroSection />
+      <InformationSection />
+      <NewsSection />
+      <PeopleSection />
+      <LicenseCitationSection />
+    </LandingLayout>
+  )
+}
